Rename to shogun
No other changes - will probably add more headers later
to note this came from datargs
