--- datargs/dispatch/concrete/__init__.py
@@ -1,3 +1,0 @@
-import datargs.dispatch.concrete.bool
-import datargs.dispatch.concrete.dataclass
-import datargs.dispatch.concrete.enum_
