+++ shogun/dispatch/concrete/__init__.py
@@ -1,0 +1,3 @@
+import shogun.dispatch.concrete.bool
+import shogun.dispatch.concrete.dataclass
+import shogun.dispatch.concrete.enum_
