+++ shogun/records/attrs.py
