+++ shogun/records/error.py
