+++ shogun/records/generic.py
