+++ shogun/utils.py
